Clase 19: Politicas de Acceso | Se crea la clase RecipePolicy y se agrega la autorizacion en el metodo update y delete, algo a tener en cuenta es que no funciona el metodo authorize() si no se trae la clase Illuminate\Foundation\Auth\Access\AuthorizesRequests y si no se usa su trait use AuthorizesRequests

+++ app/Http/Controllers/Api/RecipeController.php
@@ -11,9 +11,13 @@ use App\Http\Requests\StoreRecipeRequest;
 use App\Http\Requests\UpdateRecipeRequest;
 
 use Illuminate\Http\Request;
+use App\Policies\RecipePolicy;
+
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests; // si esto no puedo usar la policy
 
 class RecipeController extends Controller
 {
+    use AuthorizesRequests; // si esto no puedo usar la policy
     public function index()
     {
 
@@ -46,6 +50,7 @@ class RecipeController extends Controller
 
     public function update(UpdateRecipeRequest $request, Recipe $recipe)
     {
+        $this->authorize('update', $recipe);
         $recipe->update($request->all());
 
         if ($tags = json_decode($request->tags))
@@ -59,6 +64,7 @@ class RecipeController extends Controller
 
     public function destroy(Recipe $recipe)
     {
+        $this->authorize('delete', $recipe);
         $recipe->delete();
         return response()->json(null, Response::HTTP_NO_CONTENT);//204
     }
